+++ plone/app/widgets/at.py
@@ -125,6 +125,12 @@ class DateWidget(BaseWidget):
             get_date_options(request),
             args['pattern_options'])
 
+        if 'date' in args['pattern_options'] and \
+           'firstDay' in args['pattern_options']['date'] and \
+           callable(args['pattern_options']['date']['firstDay']):
+            args['pattern_options']['date']['firstDay'] = \
+                args['pattern_options']['date']['firstDay']()
+
         return args
 
     security = ClassSecurityInfo()
@@ -133,12 +139,14 @@ class DateWidget(BaseWidget):
     def process_form(self, instance, field, form, empty_marker=None):
         """Basic impl for form processing in a widget"""
 
-        value = form.get(field.getName(), None)
-        if not value:
+        value = form.get(field.getName(), empty_marker)
+        if value is empty_marker:
             return empty_marker
 
+        value = value.split('-')
+
         try:
-            value = DateTime(datetime(*map(int, value.split('-'))))
+            value = DateTime(datetime(*map(int, value)))
         except:
             return empty_marker
 
@@ -219,20 +227,23 @@ class DatetimeWidget(DateWidget):
     def process_form(self, instance, field, form, empty_marker=None):
         """Basic impl for form processing in a widget"""
 
-        value = form.get(field.getName(), None)
-        if not value:
-            return empty_marker, {}
+        value = form.get(field.getName(), empty_marker)
+        if value is empty_marker:
+            return empty_marker
 
         tmp = value.split(' ')
         if not tmp[0]:
             return empty_marker
         value = tmp[0].split('-')
-        value += tmp[1].split(':')
+        if len(tmp) == 2 and ':' in tmp[1]:
+            value += tmp[1].split(':')
+        else:
+            value += ['00', '00']
 
         try:
             value = DateTime(datetime(*map(int, value)))
         except:
-            return empty_marker, {}
+            return empty_marker
 
         return value, {}
 
@@ -351,6 +362,61 @@ registerWidget(
     used_for=('Products.Archetypes.Field.LinesField',))
 
 
+class KeywordsWidget(AjaxSelectWidget):
+    """Keywords widget for Archetypes."""
+
+    _base = InputWidget
+
+    _properties = BaseWidget._properties.copy()
+    _properties.update({
+        'pattern': 'select2',
+        'pattern_options': {},
+        'separator': ';',
+        'vocabulary': 'plone.app.vocabularies.Keywords',
+        'vocabulary_view': '@@getVocabulary',
+    })
+
+    def _base_args(self, context, field, request):
+        args = super(KeywordsWidget, self)._base_args(context, field,
+                                                      request)
+
+        membership = getToolByName(context, 'portal_membership')
+        user = membership.getAuthenticatedMember()
+
+        site_properties = getToolByName(
+            context, 'portal_properties')['site_properties']
+        allowRolesToAddKeywords = site_properties.getProperty(
+            'allowRolesToAddKeywords', None)
+
+        allowNewItems = False
+        if allowRolesToAddKeywords and [
+            role for role in user.getRolesInContext(context)
+                if role in allowRolesToAddKeywords]:
+            allowNewItems = True
+
+        args.setdefault('pattern_options', {})
+        args['pattern_options']['allowNewItems'] = allowNewItems
+
+        return args
+
+    security = ClassSecurityInfo()
+    security.declarePublic('process_form')
+
+    def process_form(self, instance, field, form, empty_marker=None):
+        value = form.get(field.getName(), empty_marker)
+        if value is empty_marker:
+            return empty_marker
+        value = value.strip().split(self.separator)
+        return value, {}
+
+
+registerWidget(
+    KeywordsWidget,
+    title='Keywords widget',
+    description=('Keywords widget'),
+    used_for=('Products.Archetypes.Field.LinesField',))
+
+
 class RelatedItemsWidget(BaseWidget):
     """Related items widget for Archetypes."""
 
@@ -363,6 +429,7 @@ class RelatedItemsWidget(BaseWidget):
         'separator': ';',
         'vocabulary': 'plone.app.vocabularies.Catalog',
         'vocabulary_view': '@@getVocabulary',
+        'allow_sorting': True,
     })
 
     def _base_args(self, context, field, request):
@@ -384,6 +451,7 @@ class RelatedItemsWidget(BaseWidget):
         args['value'] = self.separator.join(value)
 
         args.setdefault('pattern_options', {})
+        args['pattern_options']['orderable'] = self.allow_sorting
         args['pattern_options'] = dict_merge(
             get_relateditems_options(context, args['value'], self.separator,
                                      self.vocabulary, self.vocabulary_view),